[e2e]reorg zipkin tests
