+++ tests/integration/telemetry/tracing/zipkin/server_tracing_test.go
@@ -15,7 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package client
+package zipkin
 
 import (
 	"errors"
@@ -23,66 +23,41 @@ import (
 	"testing"
 	"time"
 
-	"github.com/google/uuid"
-
 	"istio.io/istio/pkg/test/framework"
-	"istio.io/istio/pkg/test/framework/components/istio"
-	"istio.io/istio/pkg/test/framework/label"
-	"istio.io/istio/pkg/test/framework/resource"
 	"istio.io/istio/pkg/test/util/retry"
 	"istio.io/istio/tests/integration/telemetry/tracing"
 )
 
-// TestClientTracing exercises the trace generation features of Istio, based on the Envoy Trace driver for zipkin using
-// client initiated tracing using envoy traceheader.
+// TestServerTracing exercises the trace generation features of Istio, based on the Envoy Trace driver for zipkin.
 // The test verifies that all expected spans (a client span and a server span for each service call in the sample bookinfo app)
 // are generated and that they are all a part of the same distributed trace with correct hierarchy and name.
-func TestClientTracing(t *testing.T) {
+//
+// More information on distributed tracing can be found here: https://istio.io/docs/tasks/telemetry/distributed-tracing/zipkin/
+func TestServerTracing(t *testing.T) {
 	framework.NewTest(t).
-		Features("observability.telemetry.tracing.client").
+		Features("observability.telemetry.tracing.server").
 		Run(func(t framework.TestContext) {
 			appNsInst := tracing.GetAppNamespace()
 			// TODO fix tracing tests in multi-network https://github.com/istio/istio/issues/28890
 			for _, cluster := range t.Clusters().ByNetwork()[t.Clusters().Default().NetworkName()] {
-				cluster := cluster
-				t.NewSubTest(cluster.StableName()).Run(func(ctx framework.TestContext) {
-					retry.UntilSuccessOrFail(ctx, func() error {
-						// Send test traffic with a trace header.
-						id := uuid.NewString()
-						extraHeader := map[string][]string{
-							tracing.TraceHeader: {id},
-						}
-						err := tracing.SendTraffic(ctx, extraHeader, cluster)
+				t.NewSubTest(cluster.StableName()).Run(func(t framework.TestContext) {
+					retry.UntilSuccessOrFail(t, func() error {
+						err := tracing.SendTraffic(t, nil, cluster)
 						if err != nil {
 							return fmt.Errorf("cannot send traffic from cluster %s: %v", cluster.Name(), err)
 						}
-						traces, err := tracing.GetZipkinInstance().QueryTraces(100,
+
+						traces, err := tracing.GetZipkinInstance().QueryTraces(300,
 							fmt.Sprintf("server.%s.svc.cluster.local:80/*", appNsInst.Name()), "")
 						if err != nil {
 							return fmt.Errorf("cannot get traces from zipkin: %v", err)
 						}
-						if !tracing.VerifyEchoTraces(ctx, appNsInst.Name(), cluster.Name(), traces) {
+						if !tracing.VerifyEchoTraces(t, appNsInst.Name(), cluster.Name(), traces) {
 							return errors.New("cannot find expected traces")
 						}
 						return nil
 					}, retry.Delay(3*time.Second), retry.Timeout(80*time.Second))
 				})
-
 			}
 		})
-}
-
-func TestMain(m *testing.M) {
-	framework.NewSuite(m).
-		Label(label.CustomSetup).
-		Setup(istio.Setup(tracing.GetIstioInstance(), setupConfig)).
-		Setup(tracing.TestSetup).
-		Run()
-}
-
-func setupConfig(ctx resource.Context, cfg *istio.Config) {
-	if cfg == nil {
-		return
-	}
-	cfg.Values["meshConfig.enableTracing"] = "true"
 }
